+++ Source/CesiumRuntime/Private/CesiumPropertyArrayBlueprintLibrary.cpp
@@ -45,15 +45,15 @@ FCesiumMetadataValue UCesiumPropertyArrayBlueprintLibrary::GetValue(
 }
 
 FString UCesiumPropertyArrayBlueprintLibrary::ToString(
-    UPARAM(ref) const FCesiumPropertyArray& Array) {
+    UPARAM(ref) const FCesiumPropertyArray& Array,
+    const FString& DefaultValue) {
   TArray<FString> results;
-
   const int64 size = UCesiumPropertyArrayBlueprintLibrary::GetArraySize(Array);
   for (int64 i = 0; i < size; i++) {
     FCesiumMetadataValue value =
         UCesiumPropertyArrayBlueprintLibrary::GetValue(Array, i);
     results.Add(
-        UCesiumMetadataValueBlueprintLibrary::GetString(value, FString()));
+        UCesiumMetadataValueBlueprintLibrary::GetString(value, DefaultValue));
   }
 
   return "[" + FString::Join(results, TEXT(", ")) + "]";