+++ Source/CesiumRuntime/Public/CesiumPropertyArrayBlueprintLibrary.h
@@ -77,13 +77,16 @@ public:
    * Prints the contents of the array to a human-readable string in the format
    * "[A, B, C, ... Z]".
    * @param Array The array.
+   * @param DefaultValue The default value for any invalid elements.
    * @return A string capturing the contents of the array.
    */
   UFUNCTION(
       BlueprintCallable,
       BlueprintPure,
       Category = "Cesium|Metadata|PropertyArray")
-  static FString ToString(UPARAM(ref) const FCesiumPropertyArray& Array);
+  static FString ToString(
+      UPARAM(ref) const FCesiumPropertyArray& Array,
+      const FString& DefaultValue);
 
   PRAGMA_DISABLE_DEPRECATION_WARNINGS
   /**